requirements.txt
# Neo4j
neo4j>=5.0.0
py2neo>=2021.2.3

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.1.0

# Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.1.0

# LLM APIs (free tiers)
openai>=1.0.0
huggingface-hub>=0.19.0

# UI
streamlit>=1.29.0
plotly>=5.18.0
networkx>=3.2.0
pyvis>=0.3.2

# Utils
python-dotenv>=1.0.0
pandas>=2.1.0
pydantic>=2.5.0
requests>=2.31.0
pytest>=7.4.0
